refactor: Move test_trial_store into own module
